pyproject.toml
[project]
name = "legal-pipeline"
version = "0.1.0"
description = "Legal cases scraping and transformation pipeline"
authors = [
    {name = "Charbel Saliba",email = "charbel_cg@outlook.com"}
]
requires-python = ">=3.12"
dependencies = [
    "scrapy (>=2.14.2,<3.0.0)",
    "ipython (>=9.12.0,<10.0.0)",
    "pymongo (>=4.6,<5.0)",
    "minio (>=7.2,<8.0)",
    "beautifulsoup4 (>=4.12,<5.0)",
    "python-dotenv (>=1.0,<2.0)",
    "dagster (>=1.8,<2.0)",
    "dagster-webserver (>=1.8,<2.0)",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2,<2.0.0",
    "ruff>=0.12.12,<0.13.0",
    "pytest>=8.0,<9.0",
]

[tool.poetry]
package-mode = false

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "config", "orchestrator", "scrapy_project", "scripts", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
files = ["app", "config", "orchestrator", "scrapy_project/legal_scraper"]
explicit_package_bases = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "scrapy.*",
    "itemadapter.*",
    "pymongo.*",
    "minio.*",
    "bs4.*",
    "dotenv.*",
    "dagster.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require running services)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
